Abstract more methods to poets, will likely ditch poet model soon
